Introduce tokens as int-id amended Identifiers to the ast
o Provide new planner with PlanContext
o Move {Label|PropertyKey|RelType}Id to ast
o Sketch out token resolution in Planner
o Use PropertyKeyTokens in Property, MapExpression, and schema commands
o Use LabelToken in HasLabels
o Use RelTypeToken in Patterns

+++ community/cypher/cypher-compiler-2.1/src/main/scala/org/neo4j/cypher/internal/compiler/v2_1/planner/CardinalityEstimator.scala
@@ -33,5 +33,6 @@ trait CardinalityEstimator {
   def estimateExpandRelationship(labelIds: Seq[LabelId], relationshipType: Seq[RelTypeId], dir: Direction): Int
 }
 
-case class LabelId(id:Long) extends AnyVal
-case class RelTypeId(id:Long) extends AnyVal
+case class LabelId(id: Int) extends AnyVal
+
+case class RelTypeId(id: Int) extends AnyVal
